+++ frontend/src/components/pages/Default.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useState } from "react";
+import React, { useContext } from "react";
 import { useHistory } from "react-router-dom";
 import * as Routes from "../../constants/Routes";
 import SampleContext from "../../contexts/SampleContext";
@@ -6,7 +6,6 @@ import SampleContext from "../../contexts/SampleContext";
 import Logout from "../auth/Logout";
 import RefreshCredentials from "../auth/RefreshCredentials";
 import ResetPassword from "../auth/ResetPassword";
-import routesAPIClient from "../../APIClients/RoutesAPIClient";
 
 import NavigationBar from "../common/NavigationBar";
 
@@ -24,13 +23,6 @@ const Button = ({ text, path }: ButtonProps) => {
 
 const TeamInfoDisplay = () => {
   const { teamName, numTerms, members, isActive } = useContext(SampleContext);
-
-  const [email, setEmail] = useState<string>("");
-
-  const inviteUser = async (userEmail: string) => {
-    await routesAPIClient.inviteUser(userEmail);
-  };
-
   return (
     <div>
       <h2>Team Info</h2>
@@ -43,22 +35,6 @@ const TeamInfoDisplay = () => {
         )}
       </div>
       <div>Active: {isActive ? "Yes" : "No"}</div>
-      <div>
-        <p>Invite User:</p>
-        <input
-          type="email"
-          value={email}
-          onChange={(event) => setEmail(event.target.value)}
-          placeholder="username@domain.com"
-        />
-        <button
-          onClick={() => inviteUser(email)}
-          className="btn btn-primary"
-          type="button"
-        >
-          Invite User
-        </button>
-      </div>
     </div>
   );
 };
